requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
python-multipart
psycopg2-binary